Move TokenKind and operator kinds
