+++ src/compiler/Nodes/Nodes.cs
@@ -125,33 +125,3 @@ public sealed partial class IdentifierExpression
 {
     public Semantic<ISymbol> ReferencedSymbol { get; internal set; }
 }
-
-public enum AssignmentKind
-{
-    Assign,
-    Plus,
-    Minus,
-    Mult,
-    Div,
-}
-
-public enum UnaryKind
-{
-    Identity,
-    Negate,
-    Not,
-}
-
-public enum BinaryKind
-{
-    Plus,
-    Minus,
-    Mult,
-    Div,
-    Equal,
-    NotEqual,
-    LessThan,
-    GreaterThan,
-    LessThanOrEqual,
-    GreaterThanOrEqual,
-}
