+++ src/compiler/OperatorKinds.cs
@@ -1,0 +1,31 @@
+namespace Noa.Compiler;
+
+public enum AssignmentKind
+{
+    Assign,
+    Plus,
+    Minus,
+    Mult,
+    Div,
+}
+
+public enum UnaryKind
+{
+    Identity,
+    Negate,
+    Not,
+}
+
+public enum BinaryKind
+{
+    Plus,
+    Minus,
+    Mult,
+    Div,
+    Equal,
+    NotEqual,
+    LessThan,
+    GreaterThan,
+    LessThanOrEqual,
+    GreaterThanOrEqual,
+}
